+++ CHANGELOG.md
@@ -85,4 +85,5 @@
   - can not be blank
   - the size must varies from 3 to 255 characters
 - add Valid annotation to validate request body based on value model validation rules to create and update by id routes
-- add PrePersist model event that adds createdAt date
+- add PrePersist model event that adds createdAt and updatedAt date
+- add PreUpdate model event that update updatedAt date